Cleanup project structure and consolidate gitignore files
