+++ .gitignore
@@ -1,3 +1,47 @@
-node_modules
+# Python virtual environments
 env/
 venv/
+
+# Node dependencies
+node_modules/
+
+# Build artifacts
+dist/
+dist-ssr/
+.vite/
+__pycache__/
+*.py[cod]
+*$py.class
+
+# Database
+db.sqlite3
+
+# Environment files
+.env
+*.local
+
+# Logs
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+lerna-debug.log*
+
+# Editor directories and files
+.vscode/*
+!.vscode/extensions.json
+.idea/
+.DS_Store
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw?
+
+# Django specific
+*.pot
+*.mo
+media/
+static/
