+++ frontEnd/package.json
@@ -32,6 +32,6 @@
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.7",
     "typescript": "^5.7.3",
-    "vite": "^5.3.4"
+    "vite": "^6.2.0"
   }
 }
