+++ frontEnd/src/components/Auth/ProtectedRoute.tsx
@@ -2,8 +2,8 @@
 import { ReactNode, useEffect, useState } from "react";
 import { Navigate } from "react-router-dom";
 import { jwtDecode } from "jwt-decode";
-import api from "../api";
-import { REFRESH_TOKEN, ACCESS_TOKEN } from "../constants";
+import api from "../../api";
+import { REFRESH_TOKEN, ACCESS_TOKEN } from "../../constants";
 
 interface ProtectedRouteProps {
     children: ReactNode;
