+++ frontEnd/src/components/Client/Client.tsx
@@ -1,16 +1,14 @@
 import React from "react";
-
-
 interface ClientProps {
     firstName: string;
     lastName: string;
     phoneNumber: string;
     email: string;
 }
-
 interface FormProps {
     client: ClientProps;
 }
+
 function Client({client}:FormProps) {
     
     return (
@@ -19,7 +17,6 @@ function Client({client}:FormProps) {
             <p className="note-content">{client.lastName}</p>
             <p className="note-date">{client.phoneNumber}</p>
             <p className="note-email">{client.email}</p>
-            <p className="note-client">Client ID: {client.id}</p>
         </div>
     );
 }
