+++ frontEnd/src/components/Property/Property.tsx
