--- frontEnd/src/components/TopBar.css
@@ -1,107 +1,0 @@
-/* TopBar.css */
-@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;500&display=swap');
-
-/* Main TopBar container */
-.top-bar {
-    position: fixed;
-    top: 0;
-    left: 0;
-    font-family: 'Roboto', sans-serif;
-    width: 100%;
-    height: 70px;
-    background: linear-gradient(90deg, rgba(40, 40, 43, 1) 0%, rgba(60, 60, 63, 1) 100%);
-    backdrop-filter: blur(5px);
-    color: white;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 10px 10px;;
-    box-sizing: border-box;
-    z-index: 10; /* Ensure the top bar is above the list items */
-    flex-shrink: 0;
-    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.5);
-  }
-
-  /* Top menu shape (background behind the top bar) */
-.top-menu-shape {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 520px;
-  background: rgba(40, 40, 43, 1);
-  clip-path: polygon(40% 30%, 50% 20%, 100% 20%, 100% 0, 0 0, 0 30%);
-  z-index: -1; /* Ensure this is below the top bar and sort dropdown */
-}
-
-  /* Client Info Section */
-  .client-info {
-    padding: 5px 10px;
-    /* Increase font-size for better legibility and adjust lines */
-    font-size: 20px;
-    font-weight: 400;
-    line-height: 1.5;
-    letter-spacing: 0.5px;
-    text-shadow: 0 1px 2px rgba(0, 0, 0, 0.3);
-    border-radius: 5px;
-    /* Remove excessive vertical spacing (was 85px, now 0 to stay inside top bar) */
-    margin-top: 85px;
-  }
-
-  .sort-button-container {
-    position: relative;
-    right: 25px; 
-    top: 35px;  
-    align-items: center;
-  }
-  
-  .sort-button {
-    background-color: rgb(33, 33, 36);
-    padding: 10px;
-    font-size: 15px;
-    cursor: pointer;
-    border-radius: 10px;
-    min-width: 100px; /* Set a fixed minimum width */
-    text-align: center;
-  }
-  .sort-button:hover{
-    transform: scale(1.03);
-  }
-  
-  .sort-dropdown {
-    display: none;
-    position: absolute;
-    background-color: rgba(0, 0, 0, 0.194);
-    margin-top: -8px;
-    width: 150px;
-    box-shadow: 10px 10px 10px 0px rgba(0, 0, 0, 0.2);
-    z-index: 70; /* Ensure dropdown is above other elements */
-    top: 50px;
-    right: 0;
-    border-radius: 5px;
-  }
-  
-  .sort-dropdown button {
-    background-color: rgba(82, 82, 82, 0.64);
-    color: rgb(197, 197, 197);
-    padding: 10px;
-    text-align: left;
-    border: none;
-    width: 100%;
-    box-sizing: border-box;
-    cursor: pointer;
-  }
-  
-  .sort-button-container:hover .sort-dropdown {
-    display: block;
-  }
-
-  .client-email {
-    color: rgb(84, 155, 255);
-    cursor: pointer;
-  }
-
-  .client-email:hover {
-    color: rgb(100, 170, 255);
-    text-decoration: underline;
-  }
