+++ frontEnd/src/constants.js
@@ -1,2 +1,2 @@
 export const ACCESS_TOKEN = "access";
-export const REFRESH_TOKEN = "refresj";
+export const REFRESH_TOKEN = "refresh";
