--- frontEnd/src/fonts.css
@@ -1,1 +1,0 @@
-@import url('https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css');
