--- frontEnd/src/pages/ErrorPage.jsx
@@ -1,11 +1,0 @@
-import "../App.css"
-
-
-function ErrorPage(){
-    return <div>
-        <h1>404 Page Not Found</h1>
-    </div>
-}
-
-
-export default ErrorPage
