+++ frontEnd/src/styles/components/BottomBar.css
@@ -1,0 +1,149 @@
+.bottom-menu {
+    position: fixed;
+    padding: 0 20px;
+    font-family: Inter;
+    bottom: 0;
+    left: 0;
+    width: 100%;
+    height: 120px;
+    background: rgba(40, 40, 43, 1);
+    backdrop-filter: blur(5px);
+    color: white;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-size: 20px;
+    z-index: 10;
+}
+
+.bottom-menu-container {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 90%;
+    max-width: 1200px;
+    gap: 130px; 
+}
+
+.bottom-menu-button-container {
+    display: flex;
+    gap: 30px; 
+    margin-left: auto; 
+}
+
+.action-button {
+    position: relative;
+    background-color: rgb(33, 33, 36);
+    border: none;
+    color: rgb(255, 255, 255);
+    width: 70px;
+    height: 70px;
+    font-size: 24px;
+    cursor: pointer;
+    border-radius: 15px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    transition: all 0.3s ease;
+}
+
+.action-button:hover {
+    transform: scale(1.05);
+    background-color: rgb(45, 45, 48);
+}
+
+.action-button:active {
+    transform: scale(0.95);
+    background-color: rgb(25, 25, 28);
+}
+
+.action-button.mode-button {
+    transition: all 0.3s ease, transform 0.3s ease;
+}
+
+.action-button.mode-button.rotated {
+    transform: rotate(180deg);
+}
+
+/*TODO: Add company profile card component*/
+
+.company-profile-card {
+    position: relative;
+    width: 260px;
+    padding: 12px;
+    border-radius: 10px;
+    background-color: rgba(33, 33, 36, 0.9);
+    box-shadow: 0 0 20px rgba(0, 0, 0, 0.5);
+    user-select: none;
+    box-sizing: border-box;
+    flex-shrink: 0;
+    display: flex;
+    align-items: center;
+    margin-right: auto; 
+}
+
+#company-image {
+    width: 60px;
+    height: 60px;
+    border-radius: 50%;
+    box-shadow: 0 0 10px rgba(255, 255, 255, 0.3);
+    margin-right: 15px;
+}
+
+.company-info {
+    flex-grow: 1;
+}
+
+.company-info-text {
+    font-size: 16px;
+    color: #ffffff;
+    font-family: 'Arial', sans-serif;
+}
+
+.company-name {
+    margin: 0 0 5px 0;
+    font-weight: bold;
+}
+
+.level {
+    margin: 0;
+    color: #ffd700;
+    font-weight: bold;
+}
+
+.xp-bar-container {
+    width: 100%;
+    height: 8px;
+    background-color: rgba(255, 255, 255, 0.1);
+    border-radius: 5px;
+    margin-top: 8px;
+    overflow: hidden;
+    position: relative;
+}
+
+.xp-bar {
+    width: 75%; /* Adjust this value to represent current XP progress */
+    height: 100%;
+    background: linear-gradient(90deg, #4CAF50, #8BC34A);
+    border-radius: 5px;
+    transition: width 0.5s ease;
+}
+
+.xp-glow {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: linear-gradient(90deg, transparent, rgba(255, 255, 255, 0.3), transparent);
+    animation: xp-glow-animation 2s infinite;
+}
+
+@keyframes xp-glow-animation {
+    0% {
+        transform: translateX(-100%);
+    }
+    100% {
+        transform: translateX(100%);
+    }
+}
