+++ frontEnd/src/styles/components/ClientListItem.css
@@ -1,0 +1,168 @@
+/* START OF CLIENT LIST ITEM STYLES */
+    .list-item.focused {
+        transform: scale(1.03);
+        animation: glow-blur 1s infinite linear;
+        z-index: 2;
+    }
+    
+    .list-item {
+        position: relative; 
+        font-family: Inter;
+        background: rgba(3, 3, 3, 0.212);
+        backdrop-filter: blur(25px);
+        color: white;
+        padding: 10px 20px;
+        margin: 8px 0;
+        border-radius: 10px 0 0 10px;
+        transition: transform 0.2s ease, background-color 0.3s ease;
+        box-sizing: border-box;
+        cursor: pointer;
+        z-index: 10;
+        min-width: 50%;
+        max-width: 55%;
+        transform-origin: right center;
+        user-select: none;
+        display: flex;
+        flex-direction: column;  /* Change to column layout */
+        align-items: flex-start;  /* Align items to the start */
+    }
+    
+    .list-item.focused {
+        transform: scale(1.03);
+        box-shadow: inset 0 0 5px rgba(226, 226, 226, 0.6), inset 0 0 10px rgba(192, 192, 255, 0.6), inset 0 0 20px rgba(192, 192, 255, 0.6);
+        z-index: 2;
+    }
+    
+    .list-item:hover {
+        transform: scale(1.02);
+        z-index: 2;
+    }
+    
+    .list-item-address {
+        margin-bottom: 5px;  /* Add some space between address and stars */
+    }
+    
+    .stars {
+        display: flex;
+        align-items: center;
+        margin-top: 5px;  /* Add some space above the stars */
+    }
+    
+    .star {
+        color: rgb(228, 213, 136);
+        margin-left: 2px;
+        text-shadow: 0 0 8px rgb(255, 157, 0);
+        font-size: 1.2em; 
+    }
+    
+    .list-item.focused .star {
+        color: rgb(255, 215, 0);
+        text-shadow: 0 0 12px rgb(255, 165, 0);
+    }
+    
+    .list-item.date-header{
+        font-family: 'Arial Narrow';
+        font-size: x-large;
+    }
+    
+    .list-item-header {
+        display: flex;
+        align-items: center;
+        gap: 10px;
+    }
+    
+    .profile-pic-container {
+        display: inline-block;
+        vertical-align: middle;
+    }
+    
+    .profile-pic {
+        width: 40px;
+        height: 40px;
+        border-radius: 50%;
+        object-fit: cover;
+    } 
+    
+    .address-container {
+        margin-top: 10px;
+        padding-left: 50px;
+        display: flex;
+        flex-direction: column;
+        gap: 5px;
+    }
+    
+    .address-item {
+        /* A subtle gradient adds some depth */
+        background: linear-gradient(135deg, rgba(3, 3, 3, 0.671), rgba(255, 255, 255, 0.212));
+        padding: 20px 20px;
+        border-radius: 10px 0 0 10px;
+        font-size: 14px;
+        color: white;
+        margin: -5px 0 10px;
+        min-width: 45%;
+        max-width: 60%;
+        border-left: 3px solid rgba(255, 255, 255, 0.6);
+        display: flex;
+        align-items: center;
+        box-shadow: inset 0 0 5px rgba(255, 255, 255, 0.2);
+        
+        /* Smooth transitions for background, transform, and shadow */
+        transition: background-color 0.3s ease, transform 0.3s ease, box-shadow 0.3s ease;
+        
+        /* Entry animation */
+        animation: slideInFade 0.2s ease-out forwards;
+        
+        /* Adding a subtle text shadow for a glowing feel */
+        text-shadow: 0 0 3px rgba(0, 0, 0, 0.5);
+        opacity: 0;
+        cursor: pointer;
+    }
+    
+    /* Hover effect for extra visual satisfaction */
+    .address-item:hover {
+        transform: scale(1.01);
+        box-shadow: inset 0 0 10px rgba(255, 255, 255, 0.4), 0 0 10px rgba(255, 255, 255, 0.3);
+    }
+    
+    /* Keyframes for entry animation */
+    @keyframes slideInFade {
+        0% {
+        transform: translateY(-10px);
+        opacity: 0;
+        }
+        100% {
+        transform: translateY(0);
+        opacity: 1;
+        }
+    }
+    
+    /* Optional exit animation class.
+        To use this, add the 'fade-out' class to your item before removal. */
+    .fade-out {
+        animation: slideOutFade .5s ease-out forwards;
+    }
+    
+    @keyframes slideOutFade {
+        0% {
+        transform: translateY(0);
+        opacity: 1;
+        }
+        100% {
+        transform: translateY(-10px);
+        opacity: 0;
+        }
+    }
+    
+    .gear-button {
+        background: none;
+        border: none;
+        color: #fff;
+        cursor: pointer;
+        font-size: 18px;
+        margin-left: 10px; /* some spacing from the client name */
+        padding: 5px;
+    }
+    
+    .gear-button:hover {
+        color: #ccc;
+    }
