+++ frontEnd/src/styles/components/TopBar.css
@@ -1,0 +1,94 @@
+/* TopBar.css */
+
+/* Main TopBar container */
+.top-bar {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 70px;
+  color: white;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  z-index: 10;
+}
+
+.top-menu-shape {
+  position: fixed;
+  top: 0;
+  width: 100%;
+  height: 520px;
+  background: rgba(40, 40, 43, 1);
+  clip-path: polygon(40% 30%, 50% 20%, 100% 20%, 100% 0, 0 0, 0 30%);
+  z-index: -1;
+}
+
+/* Client Info Section */
+.client-info {
+  padding: 10px 20px;
+  font-size: 20px;
+  line-height: 1.5;
+  letter-spacing: 1.1px;
+  text-shadow: 0 1px 2px rgba(0, 0, 0, 0.3);
+  margin-top: 80px;
+}
+
+.sort-button-container {
+  position: relative;
+  right: 15px; 
+  top: 30px;  
+  align-items: center;
+}
+
+.sort-button {
+  background-color: rgb(33, 33, 36);
+  padding: 10px 15px;
+  font-size: 15px;
+  cursor: pointer;
+  border-radius: 10px;
+  min-width: 100px;
+  text-align: center;
+}
+
+.sort-button:hover{
+  transform: scale(1.03);
+}
+
+.sort-dropdown {
+  display: none;
+  position: absolute;
+  background-color: rgba(0, 0, 0, 0.194);
+  margin-top: -8px;
+  width: 150px;
+  box-shadow: 10px 10px 10px 0px rgba(0, 0, 0, 0.2);
+  z-index: 70;
+  top: 50px;
+  right: 0;
+  border-radius: 5px;
+}
+
+.sort-dropdown button {
+  background-color: rgba(82, 82, 82, 0.64);
+  color: rgb(197, 197, 197);
+  padding: 10px;
+  text-align: left;
+  border: none;
+  width: 100%;
+  box-sizing: border-box;
+  cursor: pointer;
+}
+
+.sort-button-container:hover .sort-dropdown {
+  display: block;
+}
+
+.client-email {
+  color: rgb(84, 155, 255);
+  cursor: pointer;
+}
+
+.client-email:hover {
+  color: rgb(69, 153, 255);
+  text-decoration: underline;
+}
