+++ frontEnd/src/styles/pages/ErrorPage.css
@@ -1,0 +1,109 @@
+/* =======================
+    RESET & BASE STYLES
+======================= */
+* {
+    box-sizing: border-box;
+}
+
+html, body {
+    margin: 0;
+    padding: 0;
+    width: 100%;
+    height: 100%;
+    font-family: 'Poppins', Arial, sans-serif;
+}
+
+/* =======================
+    BACKGROUND ANIMATION
+======================= */
+@keyframes gradient-shift {
+    0% {
+        background-position: 0% 50%;
+    }
+    50% {
+        background-position: 100% 50%;
+    }
+    100% {
+    background-position: 0% 50%;
+    }
+}
+
+.error-container {
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    background: linear-gradient(
+        45deg,
+        #ff99ac,
+        #ff6a88,
+        #ff99ac
+    );
+    background-size: 200% 200%;
+    animation: gradient-shift 8s infinite;
+    padding-bottom: 30px;
+    color: #fff;
+}
+
+.highlight {
+    color: #ff6a88;   
+}
+
+/* =======================
+    ERROR CARD
+======================= */
+.error-card {
+    background: #ffffff;
+    color: #333;
+    padding: 2rem;
+    border-radius: 16px;
+    width: 90%;
+    max-width: 400px;
+    text-align: center;
+    margin: 0 auto;
+    box-shadow: 0 6px 18px rgba(0, 0, 0, 0.12);
+}
+
+.error-card h2 {
+    margin-bottom: 1.5rem;
+    font-size: 2rem;
+}
+
+.error-card p {
+    margin-bottom: 1rem;
+    line-height: 1.5;
+}
+
+/* =======================
+    BUTTON STYLE
+======================= */
+.auth-btn {
+    width: 100%;
+    padding: 0.8rem;
+    margin-top: 1rem;
+    font-size: 1rem;
+    font-weight: 600;
+    border: none;
+    border-radius: 8px;
+    color: #fff;
+    cursor: pointer;
+    background: linear-gradient(135deg, #ff6a88, #ff99ac);
+    transition: background 0.3s ease-in-out, transform 0.2s ease;
+}
+
+.auth-btn:hover {
+    background: linear-gradient(135deg, #ff6a88, #ff99ac);
+    transform: translateY(-2px);
+}
+
+/* =======================
+    ERROR CODE
+======================= */
+.error-code {
+    margin-bottom: 2.5rem;
+    font-size: 6rem;
+    font-weight: 800;
+    color: #ff6a88;
+    text-shadow: 2px 2px 6px rgba(0, 0, 0, 0.1);
+}
